Move anno-apply-time-mut fn to main track
