--- contrib/examples/annotate-apply-time-mutations-custom-resource/.expected/config.yaml
@@ -1,2 +1,0 @@
-testType: eval
-image: ghcr.io/kptdev/krm-functions-catalog/krm-fn-contrib/annotate-apply-time-mutations:latest
