--- contrib/examples/annotate-apply-time-mutations-custom-resource/resources.yaml
@@ -1,44 +1,0 @@
-apiVersion: fn.kpt.dev/v1alpha1
-kind: ApplyTimeMutation
-metadata:
-  name: example
-  annotations:
-    config.kubernetes.io/local-config: "true"
-spec:
-  targetRef:
-    group: iam.cnrm.cloud.google.com
-    kind: IAMPolicyMember
-    name: my-policy
-    namespace: example-namespace
-  substitutions:
-  - sourceRef:
-      group: resourcemanager.cnrm.cloud.google.com
-      kind: Project
-      name: my-project
-      namespace: example-namespace
-    sourcePath: $.status.number
-    targetPath: $.spec.member
-    token: "${project-number}"
----
-# Source Object
-# The status.number field will be populated by a controller after the object is created.
-apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
-kind: Project
-metadata:
-  name: my-project
-  namespace: example-namespace
-spec:
-  name: My Project
-  organizationRef:
-    external: "123456789012"
-  billingAccountRef:
-    external: "AAAAAA-BBBBBB-CCCCCC"
----
-# Target Object
-apiVersion: iam.cnrm.cloud.google.com/v1beta1
-kind: IAMPolicyMember
-metadata:
-  name: my-policy
-  namespace: example-namespace
-spec:
-  member: serviceAccount:service-${project-number}@container-engine-robot.iam.gserviceaccount.com
