--- contrib/examples/annotate-apply-time-mutations-inline-comment/.expected/diff.patch
@@ -1,40 +1,0 @@
-diff --git a/resources.yaml b/resources.yaml
-index 76ca35b..127f07a 100644
---- a/resources.yaml
-+++ b/resources.yaml
-@@ -5,11 +5,34 @@ metadata:
-   namespace: example-namespace
-   annotations:
-     unmodified-key: foobarbaz
-+    config.kubernetes.io/apply-time-mutation: |
-+      - sourcePath: $.status.number
-+        sourceRef:
-+          group: foo.cloud.google.com
-+          kind: RefKind
-+          name: example-name1
-+          namespace: example-namespace
-+        targetPath: $.spec.fieldList[0]
-+      - sourcePath: $.status.number2
-+        sourceRef:
-+          apiVersion: bar.cloud.google.com/v1beta1
-+          kind: OtherKind
-+          name: example-name2
-+          namespace: example-namespace
-+        targetPath: $.spec.numberMap.a
-+      - sourcePath: $.status.number
-+        sourceRef:
-+          group: resourcemanager.cnrm.cloud.google.com
-+          kind: Project
-+          name: example-name
-+          namespace: example-namespace
-+        targetPath: $.spec.stringField
-+        token: ${ref2}
- spec:
-   fieldList:
-     - 5 # apply-time-mutation: ${foo.cloud.google.com/namespaces/example-namespace/RefKind/example-name1:$.status.number}
-     - 6
-   numberMap:
-     a: 0 # apply-time-mutation: ${bar.cloud.google.com/v1beta1/namespaces/example-namespace/OtherKind/example-name2:$.status.number2}
--  stringField: placeholder # apply-time-mutation: serviceAccount:service-${resourcemanager.cnrm.cloud.google.com/namespaces/example-namespace/Project/example-name:$.status.number}@service.gserviceaccount.com
-+  stringField: serviceAccount:service-${ref2}@service.gserviceaccount.com # apply-time-mutation: serviceAccount:service-${resourcemanager.cnrm.cloud.google.com/namespaces/example-namespace/Project/example-name:$.status.number}@service.gserviceaccount.com
-   unmodifiedField: "no-diff"
