--- contrib/examples/annotate-apply-time-mutations-inline-comment/resources.yaml
@@ -1,15 +1,0 @@
-apiVersion: foo.cloud.google.com/v1beta1
-kind: MyTestKind
-metadata:
-  name: my-test-resource
-  namespace: example-namespace
-  annotations:
-    unmodified-key: foobarbaz
-spec:
-  fieldList:
-    - 5 # apply-time-mutation: ${foo.cloud.google.com/namespaces/example-namespace/RefKind/example-name1:$.status.number}
-    - 6
-  numberMap:
-    a: 0 # apply-time-mutation: ${bar.cloud.google.com/v1beta1/namespaces/example-namespace/OtherKind/example-name2:$.status.number2}
-  stringField: placeholder # apply-time-mutation: serviceAccount:service-${resourcemanager.cnrm.cloud.google.com/namespaces/example-namespace/Project/example-name:$.status.number}@service.gserviceaccount.com
-  unmodifiedField: "no-diff"
