+++ examples/annotate-apply-time-mutations-custom-resource/.expected/config.yaml
@@ -1,0 +1,2 @@
+testType: eval
+image: ghcr.io/kptdev/krm-functions-catalog/annotate-apply-time-mutations:latest
