+++ examples/annotate-apply-time-mutations-inline-comment/.expected/diff.patch
@@ -1,0 +1,41 @@
+diff --git a/resources.yaml b/resources.yaml
+index 286b720..0e14007 100644
+--- a/resources.yaml
++++ b/resources.yaml
+@@ -5,6 +5,29 @@ metadata:
+   namespace: example
+   annotations:
+     unmodified-key: foobarbaz
++    config.kubernetes.io/apply-time-mutation: |
++      - sourcePath: $.status.replicas
++        sourceRef:
++          group: example.com
++          kind: Config
++          name: cluster-config
++          namespace: example
++        targetPath: $.spec.replicas
++      - sourcePath: $.status.port
++        sourceRef:
++          apiVersion: example.com/v1beta1
++          kind: Database
++          name: my-db
++          namespace: example
++        targetPath: $.spec.template.spec.containers[0].env[0].value
++      - sourcePath: $.status.host
++        sourceRef:
++          group: example.com
++          kind: Database
++          name: my-db
++          namespace: example
++        targetPath: $.spec.template.spec.containers[0].env[1].value
++        token: ${ref2}
+ spec:
+   replicas: 1 # apply-time-mutation: ${example.com/namespaces/example/Config/cluster-config:$.status.replicas}
+   selector:
+@@ -22,5 +45,5 @@ spec:
+         - name: DB_PORT
+           value: "0" # apply-time-mutation: ${example.com/v1beta1/namespaces/example/Database/my-db:$.status.port}
+         - name: DB_URL
+-          value: placeholder # apply-time-mutation: jdbc:postgresql://${example.com/namespaces/example/Database/my-db:$.status.host}:5432/mydb
++          value: jdbc:postgresql://${ref2}:5432/mydb # apply-time-mutation: jdbc:postgresql://${example.com/namespaces/example/Database/my-db:$.status.host}:5432/mydb
+   unmodifiedField: "no-diff"
